+++ src/main/java/store/sonyk9919/api/domain/analysis/manager/FileStorageManager.java
@@ -1,0 +1,29 @@
+package store.sonyk9919.api.domain.analysis.manager;
+
+import java.io.File;
+import java.io.IOException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+import store.sonyk9919.api.global.common.dto.ErrorStatus;
+import store.sonyk9919.api.global.common.exception.CustomException;
+
+@Slf4j
+@Component
+public class FileStorageManager {
+
+    @Value("${file.upload-dir}")
+    private String UPLOAD_DIR;
+
+    public void saveFile(MultipartFile image, String fileName) {
+        try {
+            File dest = new File(UPLOAD_DIR + fileName + ".jpg");
+            image.transferTo(dest);
+            log.info("[FileStorage] Saved successfully: {}", fileName);
+        } catch (IOException e) {
+            log.error("[FileStorage] Failed to save file: {}", fileName, e);
+            throw new CustomException(ErrorStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+}
